QuickAccessParametersService is moved into services package
QuickAccessParametersService is moved into services package.
