+++ src/main/kotlin/nl/tudelft/ewi/se/ciselab/testgenie/evosuite/SettingsArguments.kt
@@ -2,7 +2,7 @@ package nl.tudelft.ewi.se.ciselab.testgenie.evosuite
 
 import nl.tudelft.ewi.se.ciselab.testgenie.services.TestGenieSettingsService
 import nl.tudelft.ewi.se.ciselab.testgenie.settings.TestGenieSettingsState
-import nl.tudelft.ewi.se.ciselab.testgenie.toolwindow.QuickAccessParametersService
+import nl.tudelft.ewi.se.ciselab.testgenie.services.QuickAccessParametersService
 
 /**
  * This class is used for constructing the necessary parameters for running evosuite
