+++ src/main/kotlin/nl/tudelft/ewi/se/ciselab/testgenie/services/QuickAccessParametersService.kt
@@ -1,9 +1,10 @@
-package nl.tudelft.ewi.se.ciselab.testgenie.toolwindow
+package nl.tudelft.ewi.se.ciselab.testgenie.services
 
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.components.PersistentStateComponent
 import com.intellij.openapi.components.State
 import com.intellij.openapi.components.Storage
+import nl.tudelft.ewi.se.ciselab.testgenie.toolwindow.QuickAccessParametersState
 
 /**
  * This class is responsible for helping to implement persistence for Quick Access Tool Menu on a sidebar.
